requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow
python-slugify==8.0.1
django-filter==23.3
psycopg2-binary>=2.9.7
python-decouple==3.8
drf-spectacular==0.26.5
